LayoutTests/platform/ios-simulator/editing/deleting/merge-paragraph-with-style-from-rule-expected.txt

You should see "hello webKit" all in blue:
| "
"
| <p>
|   <span>
|     class="someClass"
|     style="font-size: 12pt;"
|     "hello"
|   " "
|   <span>
|     class="someClass"
|     style="font-size: 12pt;"
|     "world"
| "
"
| <p>
|   <span>
|     class="someClass"
|     style="font-size: 12pt;"
|     "WebKit"
| "
"
